scaler

+++ train.py
@@ -61,7 +61,7 @@ def create_model(
 
     if fp16:
         model = model.half()
-        
+
     # model = AutoregressiveWrapper(model)
 
     if torch.cuda.device_count() > 1:
@@ -151,7 +151,7 @@ def train(
 
             x = batch['input_ids'].to(device)
             with torch.cuda.amp.autocast(enabled=fp16):
-                loss = model(x, return_loss=True)
+                loss = model(x.float(), return_loss=True)
                 std = 0
                 if torch.cuda.device_count() > 1:
                     loss = loss.mean()
